Extract member selector dialog into its own component
