+++ web-ui/src/components/member_selector/MemberSelector.css
@@ -1,7 +1,3 @@
-.member-selector-dialog .member-selector-dialog-content {
-    margin: 5rem 1rem 1rem 1rem;
-}
-
 /* Workaround for bleeding style applied to MuiCardHeader-action */
 .member-selector-card .MuiCardHeader-action {
     padding: 0;
