+++ web-ui/src/components/member_selector/member_selector_dialog/MemberSelectorDialog.css
@@ -1,0 +1,3 @@
+.member-selector-dialog .member-selector-dialog-content {
+    margin: 5rem 1rem 1rem 1rem;
+}
